pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "episodiq-ai"
version = "0.1.0a2"
description = "Pattern mining tool for agentic trajectories"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [{ name = "Slava Burmanov", email = "sb@episodiq.ai" }]
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "greenlet>=3.3.0",
    "hdbscan>=0.8.42",
    "httpx>=0.28.1",
    "joblib>=1.4.0",
    "jinja2>=3.1.6",
    "langchain-text-splitters>=1.1.1",
    "numpy>=2.4.1",
    "pgvector>=0.4.2",
    "psycopg>=3.3.2",
    "pydantic>=2.0.0",
    "rich>=14.2.0",
    "scikit-learn>=1.5.0",
    "seq2pat>=2.0.0",
    "sqlalchemy>=2.0.45",
    "structlog>=25.5.0",
    "tenacity>=9.1.4",
    "typer>=0.21.0",
    "umap-learn>=0.5.12",
    "uvicorn[standard]>=0.40.0",
]

[project.scripts]
episodiq = "episodiq.cli.main:app"

[dependency-groups]
dev = [
    "asgi-lifespan>=2.1.0",
    "build>=1.3.0",
    "python-dotenv>=1.2.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-recording>=0.13.4",
    "respx>=0.22.0",
    "ruff>=0.14.10",
    "twine>=6.2.0",
    "datasets>=4.7.0",
    "pyyaml>=6.0.3",
    "pycatch22>=0.4.5",
    "lightgbm>=4.6.0",
    "pytest-cov>=7.1.0",
    "deptry>=0.25.1",
]

[tool.deptry]
per_rule_ignores = { DEP002 = ["greenlet", "jinja2", "psycopg", "seq2pat"], DEP004 = ["dotenv"] }

[tool.hatch.build.targets.wheel]
packages = ["episodiq"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
